+++ tools/src/com.oracle.truffle.tools.profiler/src/com/oracle/truffle/tools/profiler/CPUSampler.java
@@ -497,7 +497,7 @@ public final class CPUSampler implements Closeable {
         this.stacksBinding = this.shadowStack.install(env.getInstrumenter(), combine(f, mode), mode == Mode.EXCLUDE_INLINED_ROOTS);
 
         this.samplerTask = new SamplingTimerTask();
-        this.samplerThread.schedule(samplerTask, 0, period);
+        this.samplerThread.schedule(samplerTask, delay, period);
 
     }
 
@@ -542,14 +542,8 @@ public final class CPUSampler implements Closeable {
 
     private class SamplingTimerTask extends TimerTask {
 
-        int runcount = 0;
-
         @Override
         public void run() {
-            runcount++;
-            if (runcount < delay / period) {
-                return;
-            }
             if (delaySamplingUntilNonInternalLangInit && !nonInternalLanguageContextInitialized) {
                 return;
             }