+++ vm/src/org.graalvm.component.installer/src/org/graalvm/component/installer/commands/Installer.java
@@ -32,8 +32,8 @@ import java.nio.channels.ByteChannel;
 import java.nio.file.Files;
 import java.nio.file.LinkOption;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
+import java.nio.file.attribute.PosixFileAttributeView;
 import java.nio.file.attribute.PosixFilePermission;
 import java.nio.file.attribute.PosixFilePermissions;
 import java.util.ArrayList;
@@ -53,6 +53,7 @@ import org.graalvm.component.installer.BundleConstants;
 import org.graalvm.component.installer.CommonConstants;
 import org.graalvm.component.installer.model.ComponentRegistry;
 import org.graalvm.component.installer.Feedback;
+import org.graalvm.component.installer.SystemUtils;
 import org.graalvm.component.installer.model.ComponentInfo;
 import org.graalvm.component.installer.model.Verifier;
 
@@ -192,7 +193,7 @@ public class Installer implements Closeable {
         if (BundleConstants.PATH_LICENSE.equals(n)) {
             rel = getLicenseRelativePath();
         } else {
-            rel = Paths.get(n);
+            rel = SystemUtils.fromCommonString(n);
         }
         return getInstallPath().resolve(rel);
     }
@@ -235,7 +236,7 @@ public class Installer implements Closeable {
 
     boolean validateOneEntry(Path target, ZipEntry entry) throws IOException {
         if (entry.isDirectory()) {
-            Path dirPath = installPath.resolve(entry.getName());
+            Path dirPath = installPath.resolve(SystemUtils.fromCommonString(entry.getName()));
             if (Files.exists(dirPath)) {
                 if (!Files.isDirectory(dirPath)) {
                     throw new IOException(
@@ -368,7 +369,7 @@ public class Installer implements Closeable {
         List<String> paths = new ArrayList<>(permissions.keySet());
         Collections.sort(paths);
         for (String s : paths) {
-            Path p = installPath.resolve(Paths.get(s));
+            Path p = installPath.resolve(SystemUtils.fromCommonString(s));
             if (Files.exists(p)) {
                 String permissionString = permissions.get(s);
                 Set<PosixFilePermission> perms;
@@ -378,19 +379,24 @@ public class Installer implements Closeable {
                 } else {
                     perms = DEFAULT_CHANGE_PERMISSION;
                 }
-                Files.setPosixFilePermissions(p, perms);
+                if (Files.getFileAttributeView(p, PosixFileAttributeView.class) != null) {
+                    Files.setPosixFilePermissions(p, perms);
+                }
             }
         }
     }
 
     public void createSymlinks() throws IOException {
+        if (SystemUtils.isWindows()) {
+            return;
+        }
         List<String> createdRelativeLinks = new ArrayList<>();
         try {
             List<String> paths = new ArrayList<>(symlinks.keySet());
             Collections.sort(paths);
             for (String s : paths) {
-                Path source = installPath.resolve(Paths.get(s));
-                Path target = Paths.get(symlinks.get(s));
+                Path source = installPath.resolve(SystemUtils.fromCommonString(s));
+                Path target = SystemUtils.fromCommonString(symlinks.get(s));
                 ensurePathExists(source.getParent());
                 createdRelativeLinks.add(s);
                 trackedPaths.add(s);